Refactor code for type safety and readability
Updated various methods to use `fluo::Vector` and `fluo::Point` for rotations, translations, and centers, improving code readability and maintainability. Replaced C-style casts with `static_cast` for better type safety. Changed loop variables to `size_t` to avoid signed/unsigned comparison issues. Simplified file operations by using `std::string` and `std::wstring` for file paths. Enhanced progress reporting by including both percentage and message string. Removed redundant code and improved type handling in several methods. Added methods for handling HSV color values and scaling in `VolumeData` and `MeshData` classes. Updated `compatibility.h` to handle long file names on Windows.

+++ FluoRender/FluoRender/Calculate/BasicStat.cpp
@@ -433,7 +433,7 @@ void BasicStat::Run()
 			//collect
 			for (size_t ii = 0; ii < bin; ++ii)
 			{
-				unsigned int key = ii + bminv;
+				unsigned int key = static_cast<unsigned int>(ii + bminv);
 				auto it = m_hist.find(key);
 				if (it == m_hist.end())
 					m_hist.insert(std::pair<unsigned int, unsigned int>(key, hist[ii]));
